+++ src/pages/index.tsx
@@ -2,7 +2,6 @@ import React from "react"
 import type { HeadFC, PageProps } from "gatsby"
 import Standard from "../components/standard/standard";
 import Book from "../components/book/book";
-import BookTut from "../components/book/book-tut";
 import ThemeSelector from "../components/theme-selector";
 import {index, header} from "../styles/index.module.css"
 import useLocalStorage from "use-local-storage";
@@ -16,7 +15,7 @@ const IndexPage: React.FC<PageProps> = () => {
         <ThemeSelector defaultTheme={theme} handleChange={
           (e: React.ChangeEvent<HTMLInputElement>) => setTheme(e.target.value)} />
       </header>
-      {theme.includes("book") ? <BookTut/> : <Standard/>}
+      {theme.includes("book") ? <Book/> : <Standard/>}
     </div>
   )
 }
