--- src/styles/book-tut.module.css
@@ -1,167 +1,0 @@
-.bookContainer {
-    height: 100%;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-}
-
-.tag:hover, .tagBack:hover, .leftPageTrigger:hover, .rightPageTrigger:hover {
-    cursor: pointer;
-}
-
-.tag, .tagBack
-{
-    width: 4%;
-    height: 8%;
-    top: -5%;
-    position: absolute;
-}
-
-.tag {
-    z-index: 100;
-}
-
-.tagBack {
-    z-index: -1;
-}
-
-.book {
-    position: relative;
-    width: 30%;
-    min-height: 375px;
-    min-width: 250px;
-    aspect-ratio: 1/1.5;
-    transition: transform 0.5s;
-}
-
-.paper {
-    position: absolute;
-    width: 100%;
-    height: 100%;
-    perspective: 5000px;
-}
-
-.frontCover, .backCover {
-    background-color: black;
-    position: absolute;
-    width: 100%;
-    height: 100%;
-    top: -2.5%;
-    transform-origin: left;
-    transition: transform 0.75s;
-    border-radius: 10px;
-}
-
-.front, .back {
-    background-color: white;
-    position: absolute;
-    width: 95%;
-    height: 95%;
-    top: 0;
-    left: 0;
-    transform-origin: left;
-    transition: transform 0.75s;
-    border-radius: 10px;
-}
-
-.rightPageTrigger{
-    min-width: 15%;
-    min-height: 100px;
-    position: absolute;
-    z-index: 500;
-    left: calc(100% - 10px - 10%); 
-    height: calc(100% - 20px);
-    opacity: 0.0;
-}
-
-
-.leftPageTrigger{
-    min-width: 15%;
-    min-height: 100px;
-    left: -100%;
-    position: absolute;
-    z-index: 500;
-    height: calc(100% - 20px);
-    opacity: 0.0;
-}
-
-/* folded corners */
-
-/* .rightPageTrigger:hover ~ .paper > .pageCornerForward:after{
-box-sizing: border-box;
-content: '';
-padding:2%;
-position: absolute;
-top: 0px;
-right: 0px;
-background: linear-gradient(-135deg, black 50%, whitesmoke 0) no-repeat;
-border: 1px solid black;
-border-bottom-left-radius: 10px;
-background-repeat: no-repeat;
-z-index: 10;
-}
-
-.leftPageTrigger:hover ~ .paper > .pageCornerBackward:before{
-box-sizing: border-box;
-content: '';
-padding:2%;
-position: absolute;
-top: 0px;
-right: -1px;
-background: linear-gradient(-135deg, black 50%, whitesmoke 0) no-repeat;
-border: 1px solid black;
-border-bottom-left-radius: 10px;
-background-repeat: no-repeat;
-z-index: 10;
-} */
-
-
-.rightPageTrigger:hover ~ .paper > .pageCornerForward {
-    transform: rotateY(-20deg);
-}
-
-.leftPageTrigger:hover ~ .paper > .pageCornerBackward{
-    transform: rotateY(-160deg);
-}
-
-
-
-.leftPageTrigger:hover ~ .paper > .backFrontCover {
-    transform: rotateY(-150deg);
-}
-
-
-.rightPageTrigger:hover ~ .paper >  .frontFrontCover {
-    transform: rotateY(-30deg);
-}
-
-.front, .frontCover {
-    z-index: 1;
-    backface-visibility: hidden;
-    border-left: 1px solid black;
-}
-
-.back, .backCover {
-    z-index: -1;
-    border-left: 1px solid black;
-}
-
-.frontContent, .backContent {
-    width: 100%;
-    height: 100%;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-}
-
-.backContent {
-    transform: rotateY(180deg);
-}
-
-.flipped .front,
-.flipped .back,
-.flipped .frontCover,
-.flipped .backCover
-{
-    transform: rotateY(-180deg);
-}
