+++ src/styles/book.module.css
@@ -1,96 +1,71 @@
-.table {
-    height: 100%;
-    border-radius: 10px;
-    background-color: brown;
-}
-
-.sleeve {
-    border-radius: 10px;
-    background-color: black;
-    padding: 10px;
-    width: 90%;
-    height: 90%;
-    position: relative;
-    display: flex;
-}
-
-.pageRight {
-    background-color: hsl(0, 0%, 96%);
-    height: 100%;
-    width: 50%;
-    border-radius: 10px;
-    position: relative;
-}
-
-.pageLeft {
-    background-color: hsl(0, 0%, 96%);
-    height: 100%;
-    width: 50%;
-    border-radius: 10px;
-    position: relative;
-}
-
-.pageRight:hover {
-    transform: translateX(-100%);
-    background-color: blue;
-}
-
-.rightPageTrigger:hover ~ .pageRight:after{
-content: '';
-width: 2.13em;
-height: 2.15em;
-position: absolute;
-top: 0px;
-right: 0px;
-background: linear-gradient(-135deg, transparent 1.5em, grey 0) no-repeat;
-border-bottom-left-radius: 10px;
-background-repeat: no-repeat;
-}
-
-.rightPageTrigger:hover ~ .pageRight {
-    background: linear-gradient(-135deg, transparent 1.5em, hsl(0, 0%, 96%) 0);
-}
-
-.leftPageTrigger:hover ~ .pageLeft:before{
-content: '';
-width: 2.13em;
-height: 2.15em;
-position: absolute;
-background: linear-gradient(-225deg, transparent 1.5em, grey 0) no-repeat;
-border-bottom-right-radius: 10px;
-background-repeat: no-repeat;
-}
-
-.leftPageTrigger:hover ~ .pageLeft {
-    background: linear-gradient(-225deg, transparent 1.5em, hsl(0, 0%, 96%) 0);
-}
-
-
-
 .bookContainer {
-    display: flex;
     height: 100%;
+    display: flex;
     justify-content: center;
     align-items: center;
 }
 
-.leftPageTrigger{
-    background-color: white;
-    min-width: 10%;
-    min-height: 100px;
-    position: absolute;
-    z-index: 500;
-    height: calc(100% - 20px);
-    opacity: 0.0;
+.tag:hover, .tagBack:hover, .leftPageTrigger:hover, .rightPageTrigger:hover {
+    cursor: pointer;
 }
 
-.leftPageTrigger:hover {
-    background-color: blue;
+.tag, .tagBack
+{
+    width: 4%;
+    height: 8%;
+    top: -5%;
+    position: absolute;
+}
+
+.tag {
+    z-index: 100;
+}
+
+.tagBack {
+    z-index: -1;
+}
+
+.book {
+    position: relative;
+    width: 30%;
+    min-height: 375px;
+    min-width: 250px;
+    aspect-ratio: 1/1.5;
+    transition: transform 0.5s;
+}
+
+.paper {
+    position: absolute;
+    width: 100%;
+    height: 100%;
+    perspective: 5000px;
+}
+
+.frontCover, .backCover {
+    background-color: black;
+    position: absolute;
+    width: 100%;
+    height: 100%;
+    top: -2.5%;
+    transform-origin: left;
+    transition: transform 0.75s;
+    border-radius: 10px;
+}
+
+.front, .back {
+    background-color: white;
+    position: absolute;
+    width: 95%;
+    height: 95%;
+    top: 0;
+    left: 0;
+    transform-origin: left;
+    transition: transform 0.75s;
+    border-radius: 10px;
 }
 
 .rightPageTrigger{
-    background-color: white;
-    min-width: 10%;
+    min-width: 15%;
     min-height: 100px;
     position: absolute;
     z-index: 500;
@@ -99,19 +74,94 @@ background-repeat: no-repeat;
     opacity: 0.0;
 }
 
-.rightPageTrigger:hover {
-    background-color: blue;
+
+.leftPageTrigger{
+    min-width: 15%;
+    min-height: 100px;
+    left: -100%;
+    position: absolute;
+    z-index: 500;
+    height: calc(100% - 20px);
+    opacity: 0.0;
+}
+
+/* folded corners */
+
+/* .rightPageTrigger:hover ~ .paper > .pageCornerForward:after{
+box-sizing: border-box;
+content: '';
+padding:2%;
+position: absolute;
+top: 0px;
+right: 0px;
+background: linear-gradient(-135deg, black 50%, whitesmoke 0) no-repeat;
+border: 1px solid black;
+border-bottom-left-radius: 10px;
+background-repeat: no-repeat;
+z-index: 10;
+}
+
+.leftPageTrigger:hover ~ .paper > .pageCornerBackward:before{
+box-sizing: border-box;
+content: '';
+padding:2%;
+position: absolute;
+top: 0px;
+right: -1px;
+background: linear-gradient(-135deg, black 50%, whitesmoke 0) no-repeat;
+border: 1px solid black;
+border-bottom-left-radius: 10px;
+background-repeat: no-repeat;
+z-index: 10;
+} */
+
+
+.rightPageTrigger:hover ~ .paper > .pageCornerForward {
+    transform: rotateY(-20deg);
+}
+
+.leftPageTrigger:hover ~ .paper > .pageCornerBackward{
+    transform: rotateY(-160deg);
 }
 
 
-/* Tuttorial */
+
+.leftPageTrigger:hover ~ .paper > .backFrontCover {
+    transform: rotateY(-150deg);
+}
 
 
-.paper {
-    border: 1px solid red;
-    position: absolute;
+.rightPageTrigger:hover ~ .paper >  .frontFrontCover {
+    transform: rotateY(-30deg);
+}
+
+.front, .frontCover {
+    z-index: 1;
+    backface-visibility: hidden;
+    border-left: 1px solid black;
+}
+
+.back, .backCover {
+    z-index: -1;
+    border-left: 1px solid black;
+}
+
+.frontContent, .backContent {
     width: 100%;
     height: 100%;
-    top: 0;
-    left: 0;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+}
+
+.backContent {
+    transform: rotateY(180deg);
+}
+
+.flipped .front,
+.flipped .back,
+.flipped .frontCover,
+.flipped .backCover
+{
+    transform: rotateY(-180deg);
 }
